Working on roslyn magic
