+++ docs/Docs.csproj
@@ -35,6 +35,7 @@
     <PackageReference Include="Microsoft.Playwright" Version="1.13.0-next-1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="Statiq.CodeAnalysis" Version="1.0.0-beta.56" />
+    <PackageReference Include="Statiq.Common" Version="1.0.0-beta.56" />
     <PackageReference Include="Statiq.Web" Version="1.0.0-beta.41" />
     <PackageReference Include="MinVer" PrivateAssets="All" Version="2.3.1" />
   </ItemGroup>
