+++ docs/input/_layout.cshtml
@@ -1,3 +1,5 @@
+@using Docs.Pipelines
+@using System.Configuration
 <!DOCTYPE html>
 <html lang="en">
 <head>
